refactor: 메세지를 저장하고 불러오는 커스텀 훅
